# Faster inference: Implemented EOT for causal sampling stopping