Through printing each generated token id tensor, I found token `2048` to be the EOT token.

Before, the inference code sampled all tokens despite EOTs being emitted. Therefore, I implemented early stopping in `casual.py`.