@jay2jp 2024-02-09
@SinanAkkoyun Do you know how much faster this would be ?