@SinanAkkoyun 2024-02-09
@jay2jp Yup! With a 4090 and flash devoding, 850 tokens (the amount of data that was being generated before, no matter how short the sentence) takes 8 to 9 seconds.
With this PR, when you for example generate the sentence "I would love to see a speedup like that." it will take around 2 to 3 seconds.

But a long sentence will still take 8 seconds, as I just stop generation when the sentence ends.