@pyetras 2024-02-12
Thanks Sinan, I had a brief look and I think this won't work with batching. I have this + a few other changes implemented in https://github.com/metavoiceio/metavoice-src/pull/46 so I'll close your PR :)